+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/incremental/UpToDateChecking.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 DiffPlug
+ * Copyright 2021-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,11 @@
  */
 package com.diffplug.spotless.maven.incremental;
 
+import java.io.File;
+import java.nio.file.Path;
+
+import javax.annotation.Nullable;
+
 import org.apache.maven.plugins.annotations.Parameter;
 
 public class UpToDateChecking {
@@ -22,7 +27,15 @@ public class UpToDateChecking {
 	@Parameter
 	private boolean enabled;
 
+	@Parameter
+	private String indexFile;
+
 	public boolean isEnabled() {
 		return enabled;
+	}
+
+	@Nullable
+	public Path getIndexFile() {
+		return indexFile == null ? null : new File(indexFile).toPath();
 	}
 }